+++ CHANGELOG.md
@@ -8,6 +8,7 @@ _Tested up to WordPress 5.0_
 ### Bug fixes
 
 - Make sure the file exists when trying to get its size.
+- Make sure a directory still exists on the server before deleting it.
 
 
 ---